+++ app/javascript/components/Head.jsx
@@ -1,11 +1,22 @@
+/* eslint-disable react/prefer-stateless-function */
 import React from 'react';
+import PropTypes from 'prop-types';
 
-const Head = () => (
-  <div>
-    <nav className="navbar navbar-inverse bg-blue fixed-top">
-      <h1 className="navbar-brand abs">Home</h1>
-    </nav>
-  </div>
-);
+class Head extends React.Component {
+  render() {
+    const { title } = this.props;
+    return (
+      <>
+        <nav className="navbar navbar-inverse bg-blue fixed-top">
+          <h1 className="navbar-brand abs">{title}</h1>
+        </nav>
+      </>
+    );
+  }
+}
+
+Head.propTypes = {
+  title: PropTypes.string.isRequired,
+};
 
 export default Head;
