--- app/javascript/components/addReading/Head.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Head = () => (
-  <div>
-    <nav className="navbar navbar-inverse bg-blue fixed-top">
-      <h1 className="navbar-brand abs">Add Reading</h1>
-    </nav>
-  </div>
-);
-
-export default Head;
