+++ app/javascript/components/addReading/ReadingForm.jsx
@@ -99,23 +99,7 @@ class ReadingForm extends React.Component {
 
   next() {
     let { currentStep } = this.state;
-    switch (currentStep) {
-      case 2:
-        currentStep = 3;
-        break;
-      case 3:
-        currentStep = 4;
-        break;
-      case 4:
-        currentStep = 5;
-        break;
-      case 5:
-        currentStep = 6;
-        break;
-      default:
-        currentStep += 1;
-    }
-
+    currentStep = currentStep === 6 ? 6 : currentStep + 1;
     this.setState({
       currentStep,
     });
@@ -123,15 +107,12 @@ class ReadingForm extends React.Component {
 
   prev() {
     let { currentStep } = this.state;
-    currentStep = currentStep <= 1 ? 1 : currentStep - 1;
+    currentStep = currentStep === 1 ? 1 : currentStep - 1;
     this.setState({
       currentStep,
     });
   }
 
-  /*
-  * the functions for our button
-  */
   previousButton() {
     const { currentStep } = this.state;
     if (currentStep !== 1) {
