+++ app/javascript/containers/AddReading.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Head from '../components/addReading/Head';
+import Head from '../components/Head';
 import Navbar from '../components/Navbar';
 import ReadingForm from '../components/addReading/ReadingForm';
 
 export default () => (
   <div>
-    <Head />
+    <Head title="Add Readings" />
     <ReadingForm />
     <Navbar />
   </div>
