+++ app/javascript/containers/Home.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/destructuring-assignment */
 import React from 'react';
-import Head from '../components/home/Head';
+import Head from '../components/Head';
 import Signin from '../components/home/Signin';
 import Signup from '../components/home/Signup';
 
@@ -30,7 +30,7 @@ class Home extends React.Component {
     const sector = this.state.section;
     return (
       <div>
-        <Head />
+        <Head title="Energy Tracker" />
         <div className="container text-content sign-in-up">
           <div className="row justify-content-center">
             <div className="col-md-4 col-md-offset-4">
