+++ app/javascript/containers/Readings.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import Head from '../components/readings/Head';
+import Head from '../components/Head';
 import Navbar from '../components/Navbar';
 import ReadingList from '../components/readings/ReadingList';
 
 export default () => (
   <div>
-    <Head />
+    <Head title="Track.It" />
     <ReadingList />
     <Navbar />
   </div>
